requirements.txt
# AWS SDK
boto3==1.34.0

# Vector Database
pinecone-client==3.1.0

# Environment Variables
python-dotenv==1.0.0

# HTTP Requests (for testing/API calls)
requests==2.31.0


sentence-transformers>=2.6.0
torch>=2.6.0 --index-url https://download.pytorch.org/whl/cpu
